Updated Message Sending Controller
The basic outline for sending a message is in a new method
(sendMessage). Still need to work on being able to create a message and
then redirect to immediately send it. Messed around with some front end
view files related to messaging (will probably be changed by front end
team eventually).